Updated build system for 'debug' target
Also updated .gitignore so git ignores the files the build system
creates in order to 'save' the last build state (so if we build 'debug'
twice, it doesn't do any extra work, but if we build 'debug' then
'release' it does a complete rebuild to strip all the debugging
information).  I'm pretty proud of the new build system, although it
still has one or two little niggles that I need to iron out at some
point.